backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
bcrypt
requests
httpx
python-dotenv